AB#1278 Use new App6Symbol
